fix(sdk): fix generated interfaces and SortQueryInput enum
